make /facets use the same 'aggregation' code as /search
